test(web): move the report-lead suite into web/tests/
Tests of web/src code belong in web/tests/, mirroring the tested module's path
below web/src/ and named for that module, so the web CI runs them and the
module resolves natively. Moving it there drops the core-side test-all.mjs
section and the update-system.mjs SYSTEM_PATHS registration entirely: the web
suite already collects tests/**/*.test.mjs, so the hook was duplicating what
npm test in web/ does.

Living inside web/ also makes the core-only-install guard dead code, so the
existsSync check and the dynamic import collapse into a plain relative import
of the module under test. Every assertion is unchanged.
