+++ update-system.mjs
@@ -346,7 +346,6 @@ const SYSTEM_PATHS = [
   'playwright.cv.config.mjs',
   'test/cv-templates.test.mjs',
   'test/cover-resolver.test.mjs',
-  'test/report-lead-section.test.mjs',
   'test/pipeline-lock.test.mjs',
   'test/profile-photo.test.mjs',
   'templates/cv-template.zh-minimal.html',
